openonload-201205-u1
    8edc22de205ffb4604beb12858da1ec4  openonload-201205-u1.tgz

+++ src/driver/linux_onload/sigaction_calls.c
@@ -352,7 +352,7 @@ efab_signal_get_tramp_data(struct mm_signal_data *tramp_data)
   return 0;
 }
 
-asmlinkage int
+asmlinkage long
 efab_linux_trampoline_sigaction(int sig, const struct sigaction *act,
                                 struct sigaction *oact, size_t sigsetsize)
 {
